CI fix round 6: fix all -Wconversion narrowing errors (49 total across 7 files)
Fixes all implicit narrowing conversions caught by GCC -Wconversion:
- net.c: uint32_t→uint8_t IP bytes, uint16_t→uint8_t ihl, size_t→uint16_t payload
- procfs.c: int→char digit/hex conversions, uint32_t→uint8_t port/IP bytes
- virtio_blk.c: int→uint16_t port I/O args
- mce_inject.c: uint64_t→uint32_t MSR address in inject loop
- mce.c: uint64_t→uint32_t in rdmsr/wrmsr calls
- nvme.c: int→uint16_t queue index, int→uint8_t IRQ vector, int→uint8_t class
- pic.c: int→uint8_t in outb mask
- vga.c: int→uint16_t cursor pos, int→uint8_t cell color
- e1000.c: uint32_t→uint8_t MAC address bytes
- pmm.c: int→uint8_t bitmap operations
- kpti.h: forward-declare struct process, fix KPTI_TRAMPOLINE_VADDR redef
- controllers.c: add stdlib.h for atoi()
- sysctl.c: int→char conversions
- splash.c: float→int conversion
- checkpoint.c: add vfs_open declaration in fs.h
- ci.yml: add -Ibuild to debug CFLAGS for generated headers

+++ src/drivers/e1000.c
@@ -321,12 +321,12 @@ static void e1000_itr_adaptive(void) {
 static void e1000_read_mac(void) {
     uint32_t ral = e1000_read(REG_RAL);
     uint32_t rah = e1000_read(REG_RAH);
-    mac_addr[0] = ral & 0xFF;
-    mac_addr[1] = (ral >> 8) & 0xFF;
-    mac_addr[2] = (ral >> 16) & 0xFF;
-    mac_addr[3] = (ral >> 24) & 0xFF;
-    mac_addr[4] = rah & 0xFF;
-    mac_addr[5] = (rah >> 8) & 0xFF;
+    mac_addr[0] = (uint8_t)(ral & 0xFF);
+    mac_addr[1] = (uint8_t)((ral >> 8) & 0xFF);
+    mac_addr[2] = (uint8_t)((ral >> 16) & 0xFF);
+    mac_addr[3] = (uint8_t)((ral >> 24) & 0xFF);
+    mac_addr[4] = (uint8_t)(rah & 0xFF);
+    mac_addr[5] = (uint8_t)((rah >> 8) & 0xFF);
 }
 
 /* ── Per-queue RX/TX initialization ────────────────────────────────── */
